python fix. new python-generate target

+++ misc/python/install_pyplusplus.sh
@@ -12,24 +12,28 @@ pgx_arc="pygccxml-${version}.zip"
 ppp_arc="pyplusplus-${version}.zip"
 prefix=${PWD}/../auxiliary
 
-for file in $pgx_arc $ppp_arc
-do
-  # Download archives if missing
-  if [ ! -e $file ]
-  then
-    ${wget_exe} ${mirror}/${file}
-  fi
-  # Unpack and install
-  if [ -e $file ]
-  then
-    unzip -q $file
-    if [ -e "Py++-${version}" ]
+# Perhaps we can make a safer check in Python?
+if [ ! -d "$prefix/lib" ]
+then
+  for file in $pgx_arc $ppp_arc
+  do
+    # Download archives if missing
+    if [ ! -e $file ]
     then
-      mv Py++-${version} `basename ${ppp_arc} .zip`
+      ${wget_exe} ${mirror}/${file}
     fi
-    cd `basename ${file} .zip`
-    ${python_exe} setup.py install --prefix=${prefix}
-    cd ..
-    rm -fR `basename ${file} .zip`
-  fi
-done
+    # Unpack and install
+    if [ -e $file ]
+    then
+      unzip -q $file
+      if [ -e "Py++-${version}" ]
+      then
+        mv Py++-${version} `basename ${ppp_arc} .zip`
+      fi
+      cd `basename ${file} .zip`
+      ${python_exe} setup.py install --prefix=${prefix}
+      cd ..
+      rm -fR `basename ${file} .zip`
+    fi
+  done
+fi
